+++ GenumKit/Parsers/AssetsCatalogParser.swift
@@ -12,7 +12,8 @@ public final class AssetsCatalogParser {
 
   public init() {}
 
-  public func addImageName(name: String) -> Bool {
+  @discardableResult
+  public func addImage(named name: String) -> Bool {
     if (entries.contains { $0.name == name }) {
       return false
     } else {
@@ -21,16 +22,16 @@ public final class AssetsCatalogParser {
     }
   }
 
-  public func parseCatalog(path: String) {
-    guard let items = loadAssetCatalogContents(path) else { return }
+  public func parseCatalog(at path: String) {
+    guard let items = loadAssetCatalog(at: path) else { return }
 
     // process recursively
-    entries = process(items)
+    entries = process(items: items)
   }
 
   struct Entry {
     var name: String
-    var value: String?
+    var value: String
     var items: [Entry]?
 
     init(name: String, value: String) {
@@ -41,6 +42,7 @@ public final class AssetsCatalogParser {
     init(name: String, items: [Entry]) {
       self.name = name
       self.items = items
+      self.value = ""
     }
   }
 }
@@ -57,7 +59,7 @@ private enum AssetCatalog: String {
 extension AssetsCatalogParser {
   static let imageSetExtension = "imageset"
 
-  private func process(items: [[String: AnyObject]], prefix: String = "") -> [Entry] {
+  fileprivate func process(items: [[String: Any]], withPrefix prefix: String = "") -> [Entry] {
     var result = [Entry]()
 
     for item in items {
@@ -73,11 +75,12 @@ extension AssetsCatalogParser {
         // this is a group/folder
         let children = item[AssetCatalog.children.rawValue] as? [[String: AnyObject]] ?? []
 
-        if let providesNamespace = item[AssetCatalog.providesNamespace.rawValue] as? NSNumber where providesNamespace.boolValue {
-          let processed = process(children, prefix: "\(prefix)\(filename)/")
+        if let providesNamespace = item[AssetCatalog.providesNamespace.rawValue] as? NSNumber,
+          providesNamespace.boolValue {
+          let processed = process(items: children, withPrefix: "\(prefix)\(filename)/")
           result += [Entry(name: filename, items: processed)]
         } else {
-          let processed = process(children, prefix: prefix)
+          let processed = process(items: children, withPrefix: prefix)
           result += [Entry(name: filename, items: processed)]
         }
       }
@@ -90,12 +93,13 @@ extension AssetsCatalogParser {
 // MARK: - ACTool
 
 extension AssetsCatalogParser {
-  private func loadAssetCatalogContents(path: String) -> [[String: AnyObject]]? {
+  fileprivate func loadAssetCatalog(at path: String) -> [[String: AnyObject]]? {
     let command = Command("xcrun", arguments: "actool", "--print-contents", path)
-    let output = command.execute()
+    let output = command.execute() as Data
 
     // try to parse plist
-    guard let plist = try? NSPropertyListSerialization.propertyListWithData(output, options: .Immutable, format: nil) else { return nil }
+    guard let plist = try? PropertyListSerialization
+        .propertyList(from: output, format: nil) else { return nil }
 
     // get first parsed catalog
     guard let contents = plist as? [String: AnyObject],