+++ Pods/Commander/Sources/ArgumentParser.swift
@@ -1,38 +1,38 @@
-private enum Arg: CustomStringConvertible {
+private enum Arg : CustomStringConvertible {
   /// A positional argument
-  case Argument(String)
+  case argument(String)
 
   /// A boolean like option, `--version`, `--help`, `--no-clean`.
-  case Option(String)
+  case option(String)
 
   /// A flag
-  case Flag(Set<Character>)
+  case flag(Set<Character>)
 
-  var description: String {
+  var description:String {
     switch self {
-    case .Argument(let value):
+    case .argument(let value):
       return value
-    case .Option(let key):
+    case .option(let key):
       return "--\(key)"
-    case .Flag(let flags):
+    case .flag(let flags):
       return "-\(String(flags))"
     }
   }
 
-  var type: String {
+  var type:String {
     switch self {
-    case .Argument:
+    case .argument:
       return "argument"
-    case .Option:
+    case .option:
       return "option"
-    case .Flag:
+    case .flag:
       return "flag"
     }
   }
 }
 
 
-public struct ArgumentParserError: ErrorType, Equatable, CustomStringConvertible {
+public struct ArgumentParserError : Error, Equatable, CustomStringConvertible {
   public let description: String
 
   public init(_ description: String) {
@@ -46,24 +46,24 @@ public func ==(lhs: ArgumentParserError, rhs: ArgumentParserError) -> Bool {
 }
 
 
-public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible {
-  private var arguments: [Arg]
+public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible {
+  fileprivate var arguments:[Arg]
 
   /// Initialises the ArgumentParser with an array of arguments
   public init(arguments: [String]) {
     self.arguments = arguments.map { argument in
       if argument.characters.first == "-" {
-        let flags = argument[argument.startIndex.successor()..<argument.endIndex]
+        let flags = argument[argument.characters.index(after: argument.startIndex)..<argument.endIndex]
 
         if flags.characters.first == "-" {
-          let option = flags[flags.startIndex.successor()..<flags.endIndex]
-          return .Option(option)
+          let option = flags[flags.characters.index(after: flags.startIndex)..<flags.endIndex]
+          return .option(option)
         }
 
-        return .Flag(Set(flags.characters))
+        return .flag(Set(flags.characters))
       }
 
-      return .Argument(argument)
+      return .argument(argument)
     }
   }
 
@@ -71,25 +71,25 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
     arguments = parser.arguments
   }
 
-  public var description: String {
-    return arguments.map { $0.description }.joinWithSeparator(" ")
+  public var description:String {
+    return arguments.map { $0.description }.joined(separator: " ")
   }
 
-  public var isEmpty: Bool {
+  public var isEmpty:Bool {
     return arguments.isEmpty
   }
 
-  public var remainder: [String] {
+  public var remainder:[String] {
     return arguments.map { $0.description }
   }
 
   /// Returns the first positional argument in the remaining arguments.
   /// This will remove the argument from the remaining arguments.
   public func shift() -> String? {
-    for (index, argument) in arguments.enumerate() {
+    for (index, argument) in arguments.enumerated() {
       switch argument {
-      case .Argument(let value):
-        arguments.removeAtIndex(index)
+      case .argument(let value):
+        arguments.remove(at: index)
         return value
       default:
         continue
@@ -100,18 +100,18 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
   }
 
   /// Returns the value for an option (--name Kyle, --name=Kyle)
-  public func shiftValueForOption(name: String) throws -> String? {
+  public func shiftValueForOption(_ name:String) throws -> String? {
     return try shiftValuesForOption(name)?.first
   }
 
   /// Returns the value for an option (--name Kyle, --name=Kyle)
-  public func shiftValuesForOption(name: String, count: Int = 1) throws -> [String]? {
+  public func shiftValuesForOption(_ name:String, count:Int = 1) throws -> [String]? {
     var index = 0
     var hasOption = false
 
     for argument in arguments {
       switch argument {
-      case .Option(let option):
+      case .option(let option):
         if option == name {
           hasOption = true
           break
@@ -127,19 +127,19 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
     }
 
     if hasOption {
-      arguments.removeAtIndex(index)  // Pop option
+      arguments.remove(at: index)  // Pop option
       return try (0..<count).map { i in
         if arguments.count > index {
-          let argument = arguments.removeAtIndex(index)
+          let argument = arguments.remove(at: index)
           switch argument {
-          case .Argument(let value):
+          case .argument(let value):
             return value
           default:
             throw ArgumentParserError("Unexpected \(argument.type) `\(argument)` as a value for `--\(name)`")
           }
         }
 
-        throw ArgumentError.MissingValue(argument: "--\(name)")
+        throw ArgumentError.missingValue(argument: "--\(name)")
       }
     }
 
@@ -147,13 +147,13 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
   }
 
   /// Returns whether an option was specified in the arguments
-  public func hasOption(name: String) -> Bool {
+  public func hasOption(_ name:String) -> Bool {
     var index = 0
     for argument in arguments {
       switch argument {
-      case .Option(let option):
+      case .option(let option):
         if option == name {
-          arguments.removeAtIndex(index)
+          arguments.remove(at: index)
           return true
         }
       default:
@@ -167,18 +167,18 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
   }
 
   /// Returns whether a flag was specified in the arguments
-  public func hasFlag(flag: Character) -> Bool {
+  public func hasFlag(_ flag:Character) -> Bool {
     var index = 0
     for argument in arguments {
       switch argument {
-      case .Flag(let option):
+      case .flag(let option):
         var options = option
         if options.contains(flag) {
           options.remove(flag)
-          arguments.removeAtIndex(index)
+          arguments.remove(at: index)
 
           if !options.isEmpty {
-            arguments.insert(.Flag(options), atIndex: index)
+            arguments.insert(.flag(options), at: index)
           }
           return true
         }
@@ -193,18 +193,18 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
   }
 
   /// Returns the value for a flag (-n Kyle)
-  public func shiftValueForFlag(flag: Character) throws -> String? {
+  public func shiftValueForFlag(_ flag:Character) throws -> String? {
     return try shiftValuesForFlag(flag)?.first
   }
 
   /// Returns the value for a flag (-n Kyle)
-  public func shiftValuesForFlag(flag: Character, count: Int = 1) throws -> [String]? {
+  public func shiftValuesForFlag(_ flag:Character, count:Int = 1) throws -> [String]? {
     var index = 0
     var hasFlag = false
 
     for argument in arguments {
       switch argument {
-      case .Flag(let flags):
+      case .flag(let flags):
         if flags.contains(flag) {
           hasFlag = true
           break
@@ -220,20 +220,20 @@ public final class ArgumentParser: ArgumentConvertible, CustomStringConvertible 
     }
 
     if hasFlag {
-      arguments.removeAtIndex(index)
+      arguments.remove(at: index)
 
       return try (0..<count).map { i in
         if arguments.count > index {
-          let argument = arguments.removeAtIndex(index)
+          let argument = arguments.remove(at: index)
           switch argument {
-          case .Argument(let value):
+          case .argument(let value):
             return value
           default:
             throw ArgumentParserError("Unexpected \(argument.type) `\(argument)` as a value for `-\(flag)`")
           }
         }
 
-        throw ArgumentError.MissingValue(argument: "-\(flag)")
+        throw ArgumentError.missingValue(argument: "-\(flag)")
       }
     }